requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
python-multipart>=0.0.9
websocket-client>=1.8.0
openai-whisper>=20240930
torch>=2.2.0
numpy>=1.26.0
silero-vad>=5.1.2
onnxruntime>=1.18.0
edge-tts>=6.1.12
python-dotenv>=1.0.1
